NO DESCRIPTION

+++ src/Libraries/CTM.Core/Domain/TradeRecord/DailyRecord.cs
@@ -82,7 +82,7 @@ namespace CTM.Core.Domain.TradeRecord
         /// <summary>
         /// 成交数量
         /// </summary>
-        public int DealVolume { get; set; }
+        public decimal DealVolume { get; set; }
 
         /// <summary>
         /// 成交金额
