+++ src/Libraries/CTM.Core/Domain/TradeRecord/DeliveryRecord.cs
@@ -59,7 +59,7 @@ namespace CTM.Core.Domain.TradeRecord
         /// <summary>
         /// 成交数量
         /// </summary>
-        public int DealVolume { get; set; }
+        public decimal DealVolume { get; set; }
 
         /// <summary>
         /// 成交金额
